+++ angular.json
@@ -41,7 +41,7 @@
               {
                 "input": "projects/storefrontapp/src/styles/lib-product-configurator.scss",
                 "bundleName": "product-configurator"
-              },              
+              },
               {
                 "input": "projects/storefrontapp/src/styles/lib-storefinder.scss",
                 "bundleName": "storefinder"
@@ -685,38 +685,38 @@
         }
       }
     },
-    "tms": {
+    "tracking": {
       "projectType": "library",
-      "root": "feature-libs/tms",
-      "sourceRoot": "feature-libs/tms/src",
+      "root": "feature-libs/tracking",
+      "sourceRoot": "feature-libs/tracking/src",
       "prefix": "cx",
       "architect": {
         "build": {
           "builder": "@angular-devkit/build-angular:ng-packagr",
           "options": {
-            "tsConfig": "feature-libs/tms/tsconfig.lib.json",
-            "project": "feature-libs/tms/ng-package.json"
+            "tsConfig": "feature-libs/tracking/tsconfig.lib.json",
+            "project": "feature-libs/tracking/ng-package.json"
           },
           "configurations": {
             "production": {
-              "tsConfig": "feature-libs/tms/tsconfig.lib.prod.json"
+              "tsConfig": "feature-libs/tracking/tsconfig.lib.prod.json"
             }
           }
         },
         "test": {
           "builder": "@angular-devkit/build-angular:karma",
           "options": {
-            "main": "feature-libs/tms/test.ts",
-            "tsConfig": "feature-libs/tms/tsconfig.spec.json",
-            "karmaConfig": "feature-libs/tms/karma.conf.js"
+            "main": "feature-libs/tracking/test.ts",
+            "tsConfig": "feature-libs/tracking/tsconfig.spec.json",
+            "karmaConfig": "feature-libs/tracking/karma.conf.js"
           }
         },
         "lint": {
           "builder": "@angular-devkit/build-angular:tslint",
           "options": {
             "tsConfig": [
-              "feature-libs/tms/tsconfig.lib.json",
-              "feature-libs/tms/tsconfig.spec.json"
+              "feature-libs/tracking/tsconfig.lib.json",
+              "feature-libs/tracking/tsconfig.spec.json"
             ],
             "exclude": ["**/node_modules/**"]
           }
