+++ ci-scripts/unit-tests-sonar.sh
@@ -101,9 +101,9 @@ if [[ -n "$coverage" ]]; then
     exit 1
 fi
 
-echo "Running unit tests and code coverage for tms"
+echo "Running unit tests and code coverage for tracking"
 exec 5>&1
-output=$(ng test tms --sourceMap --watch=false --code-coverage --browsers=ChromeHeadless | tee /dev/fd/5)
+output=$(ng test tracking --sourceMap --watch=false --code-coverage --browsers=ChromeHeadless | tee /dev/fd/5)
 coverage=$(echo $output | grep -i "does not meet global threshold" || true)
 if [[ -n "$coverage" ]]; then
     echo "Error: Tests did not meet coverage expectations"
