--- feature-libs/tms/.release-it.json
@@ -1,31 +1,0 @@
-{
-  "git": {
-    "requireCleanWorkingDir": true,
-    "requireUpstream": false,
-    "tagName": "tms-${version}",
-    "commitMessage": "Bumping tms version to ${version}",
-    "tagAnnotation": "Bumping tms version to ${version}"
-  },
-  "npm": {
-    "publishPath": "./../../dist/tms"
-  },
-  "hooks": {
-    "after:version:bump": "cd ../.. && ng build tms --prod"
-  },
-  "github": {
-    "release": true,
-    "assets": ["../../docs.tar.gz", "../../docs.zip"],
-    "releaseName": "@spartacus/tms@${version}",
-    "releaseNotes": "ts-node ../../scripts/changelog.ts --verbose --lib tms --to tms-${version}"
-  },
-  "plugins": {
-    "../../scripts/release-it/bumper.js": {
-      "out": [
-        {
-          "file": "package.json",
-          "path": ["peerDependencies.@spartacus/core"]
-        }
-      ]
-    }
-  }
-}
