--- feature-libs/tms/README.md
@@ -1,7 +1,0 @@
-# Spartacus TMS Library
-
-Spartacus' Tag Management System library offers a default support for various TMS implementations (e.g. Adobe Launch, GTM, etc.).
-
-It can be added to the existing Spartacus application by running `ng add @spartacus/tms`. For more information about Spartacus schematics, visit the [official Spartacus schematics documentation page](https://sap.github.io/spartacus-docs/schematics/).
-
-For more information, see [Spartacus](https://github.com/SAP/spartacus).
