--- feature-libs/tms/public_api.ts
@@ -1,4 +1,0 @@
-/*
- * Public API Surface of tms
- */
-export * from './core/index';
