+++ feature-libs/tracking/README.md
@@ -1,0 +1,7 @@
+# Spartacus Tracking Library
+
+Spartacus' tracking library offers a various set of features related to support support, e.g. the TMS support (including the support for Adobe Launch, GTM, etc.).
+
+It can be added to the existing Spartacus application by running `ng add @spartacus/tracking`. For more information about Spartacus schematics, visit the [official Spartacus schematics documentation page](https://sap.github.io/spartacus-docs/schematics/).
+
+For more information, see [Spartacus](https://github.com/SAP/spartacus).
