+++ feature-libs/tracking/karma.conf.js
@@ -19,7 +19,7 @@ module.exports = function (config) {
     },
     reporters: ['progress', 'kjhtml', 'coverage-istanbul', 'dots'],
     coverageIstanbulReporter: {
-      dir: require('path').join(__dirname, '../../coverage/tms'),
+      dir: require('path').join(__dirname, '../../coverage/tracking'),
       reports: ['lcov', 'cobertura', 'text-summary'],
       fixWebpackSourcePaths: true,
       thresholds: {
