+++ feature-libs/tracking/ng-package.json
@@ -1,6 +1,6 @@
 {
   "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "../../dist/tms",
+  "dest": "../../dist/tracking",
   "lib": {
     "entryFile": "./public_api.ts",
     "umdModuleIds": {
