+++ feature-libs/tracking/public_api.ts
@@ -1,0 +1,4 @@
+/*
+ * Public API Surface of tracking
+ */
+export * from './tracking.module';
