+++ feature-libs/tracking/tms/public_api.ts
@@ -1,3 +1,3 @@
 export * from './config/index';
+export * from './services/index';
 export * from './tms.module';
-export * from './tms.service';
