+++ feature-libs/tracking/tracking.module.ts
@@ -1,0 +1,10 @@
+import { ModuleWithProviders, NgModule } from '@angular/core';
+
+@NgModule({})
+export class TrackingModule {
+  static forRoot(): ModuleWithProviders<TrackingModule> {
+    return {
+      ngModule: TrackingModule,
+    };
+  }
+}
