+++ projects/schematics/package.json
@@ -48,7 +48,7 @@
       "@spartacus/storefinder",
       "@spartacus/qualtrics",
       "@spartacus/cdc",
-      "@spartacus/tms",
+      "@spartacus/tracking",
       "@spartacus/setup"
     ]
   },
