+++ projects/storefrontapp/src/app/spartacus/spartacus-features.module.ts
@@ -84,7 +84,7 @@ import { CdsFeatureModule } from './features/cds-feature.module';
 import { OrderApprovalFeatureModule } from './features/order-approval-feature.module';
 import { QualtricsFeatureModule } from './features/qualtrics-feature.module';
 import { StorefinderFeatureModule } from './features/storefinder-feature.module';
-import { TmsFeatureModule } from './features/tms-feature.module';
+import { TrackingFeatureModule } from './features/tracking-feature.module';
 
 const featureModules = [];
 
@@ -97,8 +97,8 @@ if (environment.cdc) {
 if (environment.cds) {
   featureModules.push(CdsFeatureModule);
 }
-if (environment.tms) {
-  featureModules.push(TmsFeatureModule);
+if (environment.tracking) {
+  featureModules.push(TrackingFeatureModule);
 }
 
 @NgModule({
