+++ projects/storefrontapp/src/environments/environment.prod.ts
@@ -9,5 +9,5 @@ export const environment: Environment = {
   cds: build.process.env.SPARTACUS_CDS,
   b2b: build.process.env.SPARTACUS_B2B,
   cdc: build.process.env.SPARTACUS_CDC,
-  tms: build.process.env.SPARTACUS_TMS,
+  tracking: build.process.env.SPARTACUS_TRACKING,
 };
