+++ projects/storefrontapp/src/environments/models/environment.model.ts
@@ -5,5 +5,5 @@ export interface Environment {
   b2b: boolean;
   cds: boolean;
   cdc: boolean;
-  tms: boolean;
+  tracking: boolean;
 }
