+++ projects/storefrontapp/tsconfig.app.prod.json
@@ -61,7 +61,8 @@
       "@spartacus/storefinder": ["dist/storefinder"],
       "@spartacus/storefinder/occ": ["dist/storefinder/occ"],
       "@spartacus/storefinder/root": ["dist/storefinder/root"],
-      "@spartacus/tms": ["dist/tms"],
+      "@spartacus/tracking": ["dist/tracking"],
+      "@spartacus/tracking/tms": ["dist/tracking/tms"],
       "@spartacus/cdc": ["dist/cdc"],
       "@spartacus/cds": ["dist/cds"],
       "@spartacus/assets": ["dist/assets"],
