+++ projects/storefrontapp/tsconfig.server.prod.json
@@ -64,7 +64,8 @@
       "@spartacus/storefinder": ["../../dist/storefinder"],
       "@spartacus/storefinder/occ": ["../../dist/storefinder/occ"],
       "@spartacus/storefinder/root": ["../../dist/storefinder/root"],
-      "@spartacus/tms": ["../../dist/tms"],
+      "@spartacus/tracking": ["../../dist/tracking"],
+      "@spartacus/tracking/tms": ["../../dist/tracking/tms"],
       "@spartacus/cdc": ["../../dist/cdc"],
       "@spartacus/cds": ["../../dist/cds"],
       "@spartacus/assets": ["../../dist/assets"],
