+++ scripts/packages.ts
@@ -59,7 +59,7 @@ const packageJsonPaths = [
   path.join(__dirname, '..', 'feature-libs', 'organization', 'package.json'),
   path.join(__dirname, '..', 'feature-libs', 'product', 'package.json'),
   path.join(__dirname, '..', 'feature-libs', 'storefinder', 'package.json'),
-  path.join(__dirname, '..', 'feature-libs', 'tms', 'package.json'),
+  path.join(__dirname, '..', 'feature-libs', 'tracking', 'package.json'),
   path.join(__dirname, '..', 'feature-libs', 'qualtrics', 'package.json'),
   path.join(__dirname, '..', 'integration-libs', 'cdc', 'package.json'),
 ];
