+++ sonar-project.properties
@@ -1,14 +1,14 @@
 sonar.projectKey=sap_cloud-commerce-spartacus-storefront
 sonar.projectName=Cloud commerce spartacus storefront
 
-sonar.sources=projects/assets,integration-libs/cds,projects/core,projects/storefrontstyles,projects/storefrontlib,projects/schematics,projects/incubator,projects/vendor,feature-libs/organization,feature-libs/product,feature-libs/product-configurator,feature-libs/storefinder,feature-libs/tms,integration-libs/cdc,core-libs/setup
+sonar.sources=projects/assets,integration-libs/cds,projects/core,projects/storefrontstyles,projects/storefrontlib,projects/schematics,projects/incubator,projects/vendor,feature-libs/organization,feature-libs/product,feature-libs/product-configurator,feature-libs/storefinder,feature-libs/tracking,integration-libs/cdc,core-libs/setup
 sonar.exclusions=**/node_modules/**
 
-sonar.tests=integration-libs/cds,projects/core,projects/storefrontlib,feature-libs/organization,feature-libs/product,feature-libs/product-configurator,feature-libs/storefinder,feature-libs/tms,integration-libs/cdc,core-libs/setup
+sonar.tests=integration-libs/cds,projects/core,projects/storefrontlib,feature-libs/organization,feature-libs/product,feature-libs/product-configurator,feature-libs/storefinder,feature-libs/tracking,integration-libs/cdc,core-libs/setup
 sonar.test.inclusions=**/*.spec.ts
 
-sonar.typescript.lcov.reportPaths=coverage/cds/lcov.info,coverage/core/lcov.info,coverage/storefront/lcov.info,coverage/organization/lcov.info,coverage/product/lcov.info,coverage/product-configurator/lcov.info,coverage/storefinder/lcov.info,coverage/qualtrics/lcov.info,coverage/tms/lcov.info,coverage/cdc/lcov.info,coverage/setup/lcov.info
-sonar.javascript.lcov.reportPaths=coverage/cds/lcov.info,coverage/core/lcov.info,coverage/storefront/lcov.info,coverage/organization/lcov.info,coverage/product/lcov.info,coverage/product-configurator/lcov.info,coverage/storefinder/lcov.info,coverage/qualtrics/lcov.info,coverage/tms/lcov.info,coverage/cdc/lcov.info,coverage/setup/lcov.info
+sonar.typescript.lcov.reportPaths=coverage/cds/lcov.info,coverage/core/lcov.info,coverage/storefront/lcov.info,coverage/organization/lcov.info,coverage/product/lcov.info,coverage/product-configurator/lcov.info,coverage/storefinder/lcov.info,coverage/qualtrics/lcov.info,coverage/tracking/lcov.info,coverage/cdc/lcov.info,coverage/setup/lcov.info
+sonar.javascript.lcov.reportPaths=coverage/cds/lcov.info,coverage/core/lcov.info,coverage/storefront/lcov.info,coverage/organization/lcov.info,coverage/product/lcov.info,coverage/product-configurator/lcov.info,coverage/storefinder/lcov.info,coverage/qualtrics/lcov.info,coverage/tracking/lcov.info,coverage/cdc/lcov.info,coverage/setup/lcov.info
 
 sonar.typescript.tsconfigPath=tslint.json
